Решение добавить ады

+++ empty.cpp
@@ -6,7 +6,7 @@ float Empty::mineralsMin = 1;
 float Empty::mineralsGrowSpeedMax = 0.2;
 float Empty::mineralsGrowSpeedMin = 0;
 
-Empty::Empty(uint x, uint y):Cell(x,y)
+Empty::Empty(unsigned short x, unsigned short y):Cell(x,y)
 {
     childType = Cell::empty;
     randVal = 0.00001+(rand()%10000)/10000.0;
@@ -14,7 +14,7 @@ Empty::Empty(uint x, uint y):Cell(x,y)
     recalculateLocalMineralsMax();
 }
 
-void Empty::setCoords(uint x, uint y)
+void Empty::setCoords(unsigned short x, unsigned short y)
 {
     this->x = x;
     this->y = y;
@@ -38,3 +38,4 @@ void Empty::recalculateLocalMineralsMax()
     float awayFromCenter = sqrt(powf(x-Game::worldWidth/2.0,2.0)+powf(y-Game::worldHeight/2.0,2.0));
     localMineralsMax = (mineralsMax-mineralsMin)*awayFromCenter/sqrt(powf(Game::worldWidth/2.0,2.0)+powf(Game::worldHeight/2.0,2.0));
 }
+
